replace "import cplex" with attempt_import()
